Updating controllers

+++ app/Http/Controllers/StateController.php
@@ -5,10 +5,15 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response as HttpResponse;
 
+use DateTime;
+
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
+
+use Exception;
+
 use App\State;
-use DateTime;
+
 
 class StateController extends Controller
 {
@@ -18,13 +23,13 @@ class StateController extends Controller
      * @param  int  $id
      * @return Response
      */
-    public function getState ($id = null, $status = HttpResponse::HTTP_OK) {
+    public function getState ($id = null) {
         if ( $id == null) {
             $states = State::orderBy('id', 'asc')->get();
         } else {
             $states = State::find($id);
         }
-        return response()->json($states, $status);
+        return $this->response()->array($states);
     }
     /**
     * Post state
@@ -41,10 +46,11 @@ class StateController extends Controller
         $state->created_at =  $date;
 
         try {
-            $state->save();
-            return $this->getState($state->id, $status);
+             $state->save();
+             $response = $this->getState($state->id);
+             return $response->setStatusCode($status); 
         } catch (Exceptions $e) {
-            return response()->json(['error' => $e], 400);
+             return $this->response()->errorBadRequest();
         }
     }
     /**
@@ -54,15 +60,20 @@ class StateController extends Controller
      * @return Response
      */
     public function updateState (Request $request, $id) {
-        $date = new DateTime();
-        $status = HttpResponse::HTTP_ACCEPTED;
-        $state = State::find($id);
-        $state->name = $request->input('name');
-        $state->display_name = $request->input('display_name');
-        $state->description = $request->input('description');
-        $state->updated_at =  $date;
-        $state->save();
-        return $this->getState($state->id, $status);
+        try {
+            $date = new DateTime();
+            $status = HttpResponse::HTTP_ACCEPTED;
+            $state = State::find($id);
+            $state->name = $request->input('name');
+            $state->display_name = $request->input('display_name');
+            $state->description = $request->input('description');
+            $state->updated_at =  $date;
+            $state->save();
+            $response = $this->getState($state->id);
+            return $response->setStatusCode($status);
+        } catch (Exception $e) {
+            return $this->response()->errorNotFound();
+        }
     }
     /**
      * Remove the specified resource from storage.
@@ -71,14 +82,16 @@ class StateController extends Controller
      * @return Response
      */
     public function deleteState (Request $request, $id) {
-        $status = HttpResponse::HTTP_NOT_FOUND;
         $state = State::find($id);
-        // Regular Delete
-        $state->delete(); // This will work no matter what
-        // Force Delete
-        $state->users()->sync([]); // Delete relationship data
-        $state->perms()->sync([]); // Delete relationship data
-        $state->forceDelete(); // Now force delete will work regardless of whether the pivot table has cascading delete
-        return response()->json(null, $status);
+        try {           
+             // Regular Delete
+            if(isset($state)) {
+                $state->delete(); // This will work no matter what
+                return $this->response()->noContent();
+            }
+            return $this->response()->errorNotFound();
+        } catch(Exception $e) {
+            return $this->response()->errorNotFound();
+        }
     }
 }
